Update VPR for MingW compatibility

+++ vpr/SRC/pack/pb_type_graph.cpp
@@ -13,6 +13,7 @@
 
 #include <cstdio>
 #include <cstring>
+#include <cinttypes>
 using namespace std;
 
 #include "vtr_util.h"
@@ -496,7 +497,7 @@ static void free_pb_graph(t_pb_graph_node *pb_graph_node) {
 		cur = edges_head;
 		cur_num = num_edges_head;
 		edges = (t_pb_graph_edge*) cur->data_vptr;
-		for (i = 0; i < (long) cur_num->data_vptr; i++) {
+		for (i = 0; i < (intptr_t) cur_num->data_vptr; i++) {
 			vtr::free(edges[i].input_pins);
 			vtr::free(edges[i].output_pins);
 			if (edges[i].pack_pattern_indices) {
@@ -886,7 +887,7 @@ static void alloc_and_load_complete_interc_edges(
 	cur = (vtr::t_linked_vptr*) vtr::malloc(sizeof(vtr::t_linked_vptr));
 	cur->next = num_edges_head;
 	num_edges_head = cur;
-	cur->data_vptr = (void *) ((long) in_count * out_count);
+	cur->data_vptr = (void *) ((intptr_t) in_count * out_count);
 
 	for (i_inset = 0; i_inset < num_input_sets; i_inset++) {
 		for (i_inpin = 0; i_inpin < num_input_ptrs[i_inset]; i_inpin++) {
@@ -974,7 +975,7 @@ static void alloc_and_load_direct_interc_edges(
 	cur = (vtr::t_linked_vptr*) vtr::malloc(sizeof(vtr::t_linked_vptr));
 	cur->next = num_edges_head;
 	num_edges_head = cur;
-	cur->data_vptr = (void *) ((long) num_input_ptrs[0]);
+	cur->data_vptr = (void *) ((intptr_t) num_input_ptrs[0]);
 
 	/* Reallocate memory for pins and load connections between pins and record these updates in the edges */
 	for (i = 0; i < num_input_ptrs[0]; i++) {
@@ -1036,7 +1037,7 @@ static void alloc_and_load_mux_interc_edges(t_interconnect * interconnect,
 	cur = (vtr::t_linked_vptr*) vtr::malloc(sizeof(vtr::t_linked_vptr));
 	cur->next = num_edges_head;
 	num_edges_head = cur;
-	cur->data_vptr = (void *) ((long) num_input_sets);
+	cur->data_vptr = (void *) ((intptr_t) num_input_sets);
 
 	for (i_inset = 0; i_inset < num_input_sets; i_inset++) {
 		for (i_inpin = 0; i_inpin < num_input_ptrs[i_inset]; i_inpin++) {
